okok compromise is 4000 hp, bc ive playtested and
the 2000 hp boss was so cancerous, it took me about 3 minutes to kill it. Also, basically, the player can upgrade it stats using points. Everytime it hits a shot, it gets 1 point, everytime it kills an enemy, it gets 5 points, every time it kills boss enemy, it gets 20 points, for every 100 frames it survives, it gets 1 point, and upgrades cost like 5 points, then 10 points, then 15, or smth like that.